shuke_api_autotest/requirements.txt
# box==5.3.0
# data handle
jsonpath-rw==1.4.0
PyYAML==5.3.1
python-box==4.2.3
Faker==4.0.3
# http request
requests~=2.23.0
requests-toolbelt==0.9.1


# pytest and plugins
allure-pytest==2.8.13
#pytest~=5.4.1
pytest==6.2.5
pytest-html==2.1.1
pytest-metadata==1.8.0
pytest-rerunfailures==10.1


# db operations
records==0.5.3
PyMySQL==0.9.3
redis==3.5.3
python-memcached==1.59
sshtunnel==0.4.0
# tools
click==7.1.2
retrying==1.3.3
python-dateutil==2.8.1
deepdiff~=5.0.2
sure==2.0.0
rsa==4.6
beautifulsoup4==4.9.3

# log
sentry-sdk==0.19.4

# excel
pandas==1.1.5
xlrd==1.2.0

pytest-ordering==0.6
pytest-json-report==1.2.1
pytest-xdist==2.5.0